SVG: optimize transform group creation
This patch ensures that we only create transformation groups when it is
actually required and that we re-use transform groups as much as
possible. This, combined with the other patches, reduces the number of
transform groups for the Tracemonkey paper from 2790 to 1955, thereby
making the DOM much lighter.

Furthermore, we now issue a warning when an unimplemented graphic state
is encountered. Previously, this would be ignored silently, making
debugging harder.